+++ types/affinda.d.ts
@@ -1555,6 +1555,7 @@ export declare interface Collection {
     extractor?: Extractor;
     autoValidationThreshold?: number;
     fields?: FieldGroup[];
+    fieldsLayout?: FieldsLayout;
     fieldsConfigured?: boolean;
     dateFormatPreference?: CollectionDateFormatPreference;
     /** Predict the date format from any dates in the document that is not ambiguous. */
@@ -2228,18 +2229,38 @@ export declare interface ExtractorUpdate {
 
 export declare interface Field {
     label: string;
+    /** Data point identifier */
+    dataPoint: string;
+    mandatory?: boolean;
+    autoValidationThreshold?: number;
+    showDropdown?: boolean;
+    fields?: Field[];
+}
+
+export declare interface FieldCategory {
+    enabledFields: Field[];
+    disabledFields: Field[];
+}
+
+export declare interface FieldDeprecated {
+    label: string;
     slug?: string;
     dataPoint: string;
     mandatory?: boolean;
     disabled?: boolean;
     autoValidationThreshold?: number;
     showDropdown?: boolean;
-    fields?: Field[];
+    fields?: FieldDeprecated[];
 }
 
 export declare interface FieldGroup {
     label: string;
-    fields: Field[];
+    fields: FieldDeprecated[];
+}
+
+export declare interface FieldsLayout {
+    defaultCategory: FieldCategory;
+    categories: FieldCategory[];
 }
 
 export declare type FloatAnnotation = Annotation & {
@@ -2466,8 +2487,6 @@ export declare interface JobDescriptionData {
     location?: LocationAnnotation;
     certifications?: (TextAnnotation | null)[];
     yearsExperience?: YearsExperienceAnnotation;
-    /** All of the raw text of the parsed job description, example is shortened for readability */
-    rawText?: string;
 }
 
 export declare interface JobDescriptionSearch {